[CORE] Ps: tests refactor
