+++ core/src/ps/access.rs
@@ -289,141 +289,3 @@ impl<'a> RuleMut<'a> for AccessHashMap {
         Ok(())
     }
 }
-
-#[cfg(test)]
-mod test {
-    use super::*;
-    use crate::ps::array::ParseArrayLiteral;
-    use crate::ps::build_powershell_tree;
-    use crate::ps::forward::Forward;
-    use crate::ps::integer::ParseInt;
-    use crate::ps::string::ParseString;
-
-    #[test]
-    fn test_access_string_element_from_int() {
-        let mut tree = build_powershell_tree("'abc'[0, 1]").unwrap();
-        tree.apply_mut(&mut (
-            ParseInt::default(),
-            Forward::default(),
-            ParseString::default(),
-            AccessString::default(),
-            ParseArrayLiteral::default(),
-        ))
-        .unwrap();
-
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .data()
-                .expect("Inferred type"),
-            Array(vec![Str("a".to_string()), Str("b".to_string())])
-        );
-    }
-
-    #[test]
-    fn test_access_string_element_from_negative_int() {
-        let mut tree = build_powershell_tree("'abc'[-2, -1]").unwrap();
-        tree.apply_mut(&mut (
-            ParseInt::default(),
-            Forward::default(),
-            ParseString::default(),
-            AccessString::default(),
-            ParseArrayLiteral::default(),
-        ))
-        .unwrap();
-
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .data()
-                .expect("Inferred type"),
-            Array(vec![Str("b".to_string()), Str("c".to_string())])
-        );
-    }
-
-    #[test]
-    fn test_access_string_element_from_negative_string() {
-        let mut tree = build_powershell_tree("'abc'['-2']").unwrap();
-        tree.apply_mut(&mut (
-            ParseInt::default(),
-            Forward::default(),
-            ParseString::default(),
-            AccessString::default(),
-        ))
-        .unwrap();
-
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .data()
-                .expect("Inferred type"),
-            Raw(Str("b".to_string()))
-        );
-    }
-
-    #[test]
-    fn test_access_string_element_from_string() {
-        let mut tree = build_powershell_tree("'abc'['0']").unwrap();
-        tree.apply_mut(&mut (
-            ParseInt::default(),
-            Forward::default(),
-            ParseString::default(),
-            AccessString::default(),
-        ))
-        .unwrap();
-
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .data()
-                .expect("Inferred type"),
-            Raw(Str("a".to_string()))
-        );
-    }
-
-    #[test]
-    fn test_access_string_multi_element_from_int() {
-        let mut tree = build_powershell_tree("'abc'[1, 2]").unwrap();
-        tree.apply_mut(&mut (
-            ParseInt::default(),
-            Forward::default(),
-            ParseString::default(),
-            AccessString::default(),
-            ParseArrayLiteral::default(),
-        ))
-        .unwrap();
-
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .data()
-                .expect("Inferred type"),
-            Array(vec![Str("b".to_string()), Str("c".to_string())])
-        );
-    }
-}
