+++ core/src/ps/array.rs
@@ -438,233 +438,3 @@ impl<'a> RuleMut<'a> for NewObjectArray {
         Ok(())
     }
 }
-
-#[cfg(test)]
-mod test {
-    use crate::ps::Powershell::Array;
-    use crate::ps::Value::{Num, Str};
-    use crate::ps::access::AccessString;
-    use crate::ps::array::{
-        AddArray, ComputeArrayExpr, NewObjectArray, ParseArrayLiteral, ParseRange,
-    };
-    use crate::ps::build_powershell_tree;
-    use crate::ps::forward::Forward;
-    use crate::ps::integer::{AddInt, ParseInt};
-    use crate::ps::string::ParseString;
-
-    #[test]
-    fn test_init_num_array() {
-        let mut tree = build_powershell_tree("@(1,2,3)").unwrap();
-        tree.apply_mut(&mut (
-            ParseInt::default(),
-            Forward::default(),
-            ComputeArrayExpr::default(),
-            ParseArrayLiteral::default(),
-        ))
-        .unwrap();
-
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .data()
-                .expect("Inferred type"),
-            Array(vec![Num(1), Num(2), Num(3)])
-        );
-    }
-
-    #[test]
-    fn test_init_mix_array() {
-        let mut tree = build_powershell_tree("@(1,2,'3')").unwrap();
-        tree.apply_mut(&mut (
-            ParseInt::default(),
-            ParseString::default(),
-            Forward::default(),
-            ComputeArrayExpr::default(),
-            ParseArrayLiteral::default(),
-        ))
-        .unwrap();
-
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .data()
-                .expect("Inferred type"),
-            Array(vec![Num(1), Num(2), Str("3".to_string())])
-        );
-    }
-
-    #[test]
-    fn test_init_str_array() {
-        let mut tree = build_powershell_tree("@('a','b','c')").unwrap();
-        tree.apply_mut(&mut (
-            ParseString::default(),
-            Forward::default(),
-            ComputeArrayExpr::default(),
-            ParseArrayLiteral::default(),
-        ))
-        .unwrap();
-
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .data()
-                .expect("Inferred type"),
-            Array(vec![
-                Str("a".to_string()),
-                Str("b".to_string()),
-                Str("c".to_string())
-            ])
-        );
-    }
-
-    #[test]
-    fn test_init_int_array_without_at() {
-        let mut tree = build_powershell_tree("1,2,3").unwrap();
-        tree.apply_mut(&mut (
-            ParseInt::default(),
-            Forward::default(),
-            ComputeArrayExpr::default(),
-            ParseArrayLiteral::default(),
-        ))
-        .unwrap();
-
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .data()
-                .expect("Inferred type"),
-            Array(vec![Num(1), Num(2), Num(3)])
-        );
-    }
-
-    #[test]
-    fn test_init_array_with_multi_statement() {
-        let mut tree = build_powershell_tree("@(1,2,3; 4 + 6)").unwrap();
-        tree.apply_mut(&mut (
-            ParseInt::default(),
-            AddInt::default(),
-            Forward::default(),
-            ComputeArrayExpr::default(),
-            ParseArrayLiteral::default(),
-        ))
-        .unwrap();
-
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .data()
-                .expect("Inferred type"),
-            Array(vec![Num(1), Num(2), Num(3), Num(10)])
-        );
-    }
-
-    #[test]
-    fn test_concat_array() {
-        let mut tree = build_powershell_tree("'foo'[0,1] + 'x'").unwrap();
-        tree.apply_mut(&mut (
-            ParseInt::default(),
-            AddArray::default(),
-            Forward::default(),
-            ComputeArrayExpr::default(),
-            ParseArrayLiteral::default(),
-            AccessString::default(),
-            ParseString::default(),
-        ))
-        .unwrap();
-
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .data()
-                .expect("Inferred type"),
-            Array(vec![
-                Str("f".to_string()),
-                Str("o".to_string()),
-                Str("x".to_string())
-            ])
-        );
-    }
-
-    #[test]
-    fn test_negative_range() {
-        let mut tree = build_powershell_tree("-1..-3").unwrap();
-        tree.apply_mut(&mut (
-            ParseInt::default(),
-            Forward::default(),
-            ComputeArrayExpr::default(),
-            ParseArrayLiteral::default(),
-            AccessString::default(),
-            ParseString::default(),
-            ParseRange::default(),
-        ))
-        .unwrap();
-
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .data()
-                .expect("Inferred type"),
-            Array(vec![Num(-1), Num(-2), Num(-3)])
-        );
-    }
-
-    #[test]
-    fn test_new_object_array() {
-        let mut tree = build_powershell_tree("New-Object byte[] 16").unwrap();
-        tree.apply_mut(&mut (ParseInt::default(), Forward::default()))
-            .unwrap();
-        tree.apply_mut(&mut NewObjectArray::default()).unwrap();
-
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .data()
-                .expect("Inferred data"),
-            Array(vec![Num(0); 16]),
-        );
-    }
-}
