+++ core/src/ps/bool.rs
@@ -240,276 +240,3 @@ impl<'a> RuleMut<'a> for Not {
         Ok(())
     }
 }
-
-#[cfg(test)]
-mod test {
-    use crate::ps::Powershell::Raw;
-    use crate::ps::Value::Bool;
-    use crate::ps::bool::{BoolAlgebra, Comparison, ParseBool};
-    use crate::ps::build_powershell_tree;
-    use crate::ps::forward::Forward;
-    use crate::ps::integer::ParseInt;
-    use crate::ps::string::ParseString;
-
-    #[test]
-    fn test_parse_bool_true() {
-        let mut tree = build_powershell_tree("$true").unwrap();
-        tree.apply_mut(&mut (ParseBool::default(), Forward::default()))
-            .unwrap();
-
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .data()
-                .expect("Inferred type"),
-            Raw(Bool(true))
-        );
-    }
-
-    #[test]
-    fn test_parse_bool_false() {
-        let mut tree = build_powershell_tree("$false").unwrap();
-        tree.apply_mut(&mut (ParseBool::default(), Forward::default()))
-            .unwrap();
-
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .data()
-                .expect("Inferred type"),
-            Raw(Bool(false))
-        );
-    }
-
-    #[test]
-    fn test_boolean_algebra_or() {
-        let mut tree = build_powershell_tree("$true -or $false").unwrap();
-        tree.apply_mut(&mut (
-            ParseBool::default(),
-            Forward::default(),
-            BoolAlgebra::default(),
-        ))
-        .unwrap();
-
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .data()
-                .expect("Inferred type"),
-            Raw(Bool(true))
-        );
-    }
-
-    #[test]
-    fn test_boolean_algebra_and() {
-        let mut tree = build_powershell_tree("$true -and $false").unwrap();
-        tree.apply_mut(&mut (
-            ParseBool::default(),
-            Forward::default(),
-            BoolAlgebra::default(),
-        ))
-        .unwrap();
-
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .data()
-                .expect("Inferred type"),
-            Raw(Bool(false))
-        );
-    }
-
-    #[test]
-    fn test_comparison_int_int() {
-        let mut tree = build_powershell_tree("4 -le 5").unwrap();
-        tree.apply_mut(&mut (
-            ParseInt::default(),
-            Forward::default(),
-            Comparison::default(),
-        ))
-        .unwrap();
-
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .data()
-                .expect("Inferred type"),
-            Raw(Bool(true))
-        );
-    }
-
-    #[test]
-    fn test_comparison_int_str() {
-        let mut tree = build_powershell_tree("4 -le '5'").unwrap();
-        tree.apply_mut(&mut (
-            ParseInt::default(),
-            ParseString::default(),
-            Forward::default(),
-            Comparison::default(),
-        ))
-        .unwrap();
-
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .data()
-                .expect("Inferred type"),
-            Raw(Bool(true))
-        );
-    }
-
-    #[test]
-    fn test_comparison_special_case_1() {
-        let mut tree = build_powershell_tree("'True' -eq $true").unwrap();
-        tree.apply_mut(&mut (
-            ParseBool::default(),
-            ParseString::default(),
-            Forward::default(),
-            Comparison::default(),
-        ))
-        .unwrap();
-
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .data()
-                .expect("Inferred type"),
-            Raw(Bool(true))
-        );
-    }
-
-    #[test]
-    fn test_comparison_special_case_2() {
-        let mut tree = build_powershell_tree("'False' -eq $false").unwrap();
-        tree.apply_mut(&mut (
-            ParseBool::default(),
-            ParseString::default(),
-            Forward::default(),
-            Comparison::default(),
-        ))
-        .unwrap();
-
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .data()
-                .expect("Inferred type"),
-            Raw(Bool(true))
-        );
-    }
-
-    #[test]
-    fn test_comparison_special_case_3() {
-        let mut tree = build_powershell_tree("'' -eq $true").unwrap();
-        tree.apply_mut(&mut (
-            ParseBool::default(),
-            ParseString::default(),
-            Forward::default(),
-            Comparison::default(),
-        ))
-        .unwrap();
-
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .data()
-                .expect("Inferred type"),
-            Raw(Bool(false))
-        );
-    }
-
-    #[test]
-    fn test_comparison_special_case_4() {
-        let mut tree = build_powershell_tree("'' -eq $false").unwrap();
-        tree.apply_mut(&mut (
-            ParseBool::default(),
-            ParseString::default(),
-            Forward::default(),
-            Comparison::default(),
-        ))
-        .unwrap();
-
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .data()
-                .expect("Inferred type"),
-            Raw(Bool(false))
-        );
-    }
-
-    #[test]
-    fn test_comparison_special_case_5() {
-        let mut tree = build_powershell_tree("$false -eq ''").unwrap();
-        tree.apply_mut(&mut (
-            ParseBool::default(),
-            ParseString::default(),
-            Forward::default(),
-            Comparison::default(),
-        ))
-        .unwrap();
-
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .data()
-                .expect("Inferred type"),
-            Raw(Bool(true))
-        );
-    }
-}
