+++ core/src/ps/cast.rs
@@ -264,153 +264,3 @@ impl<'a> RuleMut<'a> for CastNull {
         Ok(())
     }
 }
-
-#[cfg(test)]
-mod test {
-    use crate::ps::Powershell::{Array, Raw};
-    use crate::ps::Value::{Num, Str};
-    use crate::ps::array::ParseArrayLiteral;
-    use crate::ps::build_powershell_tree;
-    use crate::ps::cast::Cast;
-    use crate::ps::foreach::{ForEach, PSItemInferrator};
-    use crate::ps::forward::Forward;
-    use crate::ps::integer::{AddInt, ParseInt};
-    use crate::ps::string::{ConcatString, ParseString};
-    use crate::ps::typing::ParseType;
-
-    #[test]
-    fn test_cast_int_to_char() {
-        let mut tree = build_powershell_tree("[char]0x61").unwrap();
-        tree.apply_mut(&mut (
-            ParseInt::default(),
-            Forward::default(),
-            Cast::default(),
-            ParseType::default(),
-        ))
-        .unwrap();
-
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .data()
-                .expect("Inferred type"),
-            Raw(Str("a".to_string()))
-        );
-    }
-
-    #[test]
-    fn test_cast_char_to_int() {
-        let mut tree = build_powershell_tree("[int]'61'").unwrap();
-        tree.apply_mut(&mut (
-            ParseString::default(),
-            Forward::default(),
-            Cast::default(),
-            ParseType::default(),
-        ))
-        .unwrap();
-
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .data()
-                .expect("Inferred type"),
-            Raw(Num(61))
-        );
-    }
-
-    #[test]
-    fn test_cast_int_additive_to_char() {
-        let mut tree = build_powershell_tree("[char](0x61 + 3)").unwrap();
-        tree.apply_mut(&mut (
-            ParseInt::default(),
-            AddInt::default(),
-            Forward::default(),
-            Cast::default(),
-            ParseType::default(),
-        ))
-        .unwrap();
-
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .data()
-                .expect("Inferred type"),
-            Raw(Str("d".to_string()))
-        );
-    }
-
-    #[test]
-    fn test_cast_int_concat_char() {
-        let mut tree =
-            build_powershell_tree("[char]0x74 + [char]0x6f + [char]0x74 + [char]0x6f").unwrap();
-        tree.apply_mut(&mut (
-            ParseInt::default(),
-            ConcatString::default(),
-            Forward::default(),
-            Cast::default(),
-            ParseType::default(),
-        ))
-        .unwrap();
-
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .data()
-                .expect("Inferred type"),
-            Raw(Str("toto".to_string()))
-        );
-    }
-
-    #[test]
-    fn test_cast_foreach_char() {
-        let mut tree = build_powershell_tree("(0x74, 0x6f, 0x74, 0x6f) | % {[char]$_}").unwrap();
-        tree.apply_mut(&mut (
-            ParseInt::default(),
-            Forward::default(),
-            Cast::default(),
-            ForEach::default(),
-            ParseArrayLiteral::default(),
-            PSItemInferrator::default(),
-            ParseType::default(),
-        ))
-        .unwrap();
-
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .data()
-                .expect("Inferred type"),
-            Array(vec![
-                Str("t".to_string()),
-                Str("o".to_string()),
-                Str("t".to_string()),
-                Str("o".to_string())
-            ])
-        );
-    }
-}
