+++ core/src/ps/cmdlets.rs
@@ -15,7 +15,7 @@ const UNIX_CMDLETS: &str = include_str!("cmdlets/unix.txt");
 // maps lowercase cmdlet name -> original case
 static CMDLET_NAMES: OnceLock<HashMap<String, String>> = OnceLock::new();
 
-fn parse_cmdlet_names(content: &str) -> impl Iterator<Item = String> + '_ {
+pub(crate) fn parse_cmdlet_names(content: &str) -> impl Iterator<Item = String> + '_ {
     content.lines().filter_map(|line| {
         let mut columns = line.split_whitespace();
         match columns.next() {
@@ -99,56 +99,5 @@ impl<'a> RuleMut<'a> for WildcardCmdlet {
             node.set(Raw(Str(resolved)));
         }
         Ok(())
-    }
-}
-
-#[cfg(test)]
-mod test {
-    use super::*;
-    use crate::ps::build_powershell_tree;
-    use crate::ps::forward::Forward;
-    use crate::ps::linter::Linter;
-    use crate::ps::strategy::PowershellStrategy;
-
-    #[test]
-    fn test_parse_skips_malformed_lines() {
-        let content = "Alias           Add-AppPackage                                     2.0.1.0    Appx\n\
-                        hodOverride                  2.1.0.0    International\n\
-                        Function        cd..\n";
-        let names: Vec<String> = parse_cmdlet_names(content).collect();
-        assert_eq!(names, vec!["Add-AppPackage", "cd.."]);
-    }
-
-    #[test]
-    fn test_resolve_unambiguous_wildcard() {
-        assert_eq!(
-            resolve_wildcard_cmdlet("G*t-Ch*dItem"),
-            Some("Get-ChildItem".to_string())
-        );
-    }
-
-    #[test]
-    fn test_resolve_ambiguous_wildcard_stays_unresolved() {
-        assert_eq!(resolve_wildcard_cmdlet("Get-*"), None);
-    }
-
-    #[test]
-    fn test_resolve_no_wildcard_returns_none() {
-        assert_eq!(resolve_wildcard_cmdlet("Get-ChildItem"), None);
-    }
-
-    #[test]
-    fn test_wildcard_cmdlet_rule_rewrites_output() {
-        let mut tree = build_powershell_tree("g*t-ch*ditem -Path C:\\foo").unwrap();
-        tree.apply_mut_with_strategy(
-            &mut (Forward::default(), WildcardCmdlet::default()),
-            PowershellStrategy::default(),
-        )
-        .unwrap();
-
-        let mut ps_litter_view = Linter::default();
-        tree.apply(&mut ps_litter_view).unwrap();
-
-        assert_eq!(ps_litter_view.output, "Get-ChildItem -Path C:\\foo");
     }
 }
