+++ core/src/ps/encoding.rs
@@ -402,343 +402,3 @@ impl<'a> RuleMut<'a> for EncodingGetBytes {
         Ok(())
     }
 }
-
-#[cfg(test)]
-mod test {
-    use crate::ps::Powershell::{Bytes, Raw};
-    use crate::ps::Value::Str;
-    use crate::ps::array::{ComputeArrayExpr, ParseArrayLiteral};
-    use crate::ps::build_powershell_tree;
-    use crate::ps::encoding::{EncodingGetBytes, EncodingGetString, EncodingType};
-    use crate::ps::forward::Forward;
-    use crate::ps::integer::ParseInt;
-    use crate::ps::string::ParseString;
-    use crate::ps::typing::ParseType;
-
-    fn root_data(
-        tree: &crate::tree::Tree<crate::tree::HashMapStorage<crate::ps::Powershell>>,
-    ) -> crate::ps::Powershell {
-        tree.root()
-            .unwrap()
-            .child(0)
-            .unwrap()
-            .child(0)
-            .unwrap()
-            .data()
-            .expect("Inferred type")
-            .clone()
-    }
-
-    #[test]
-    fn test_decode_utf8() {
-        let mut tree =
-            build_powershell_tree("[System.Text.Encoding]::utf8.getstring(@(102, 111, 111))")
-                .unwrap();
-        tree.apply_mut(&mut (
-            Forward::default(),
-            ParseType::default(),
-            EncodingType::default(),
-            ParseInt::default(),
-            ParseArrayLiteral::default(),
-            ComputeArrayExpr::default(),
-            EncodingGetString::default(),
-        ))
-        .unwrap();
-
-        assert_eq!(root_data(&tree), Raw(Str("foo".to_string())));
-    }
-
-    #[test]
-    fn test_decode_unicode() {
-        let mut tree = build_powershell_tree(
-            "[System.Text.Encoding]::Unicode.getstring(@(102, 0, 111, 0, 111, 0))",
-        )
-        .unwrap();
-        tree.apply_mut(&mut (
-            Forward::default(),
-            ParseType::default(),
-            EncodingType::default(),
-            ParseInt::default(),
-            ParseArrayLiteral::default(),
-            ComputeArrayExpr::default(),
-            EncodingGetString::default(),
-        ))
-        .unwrap();
-
-        assert_eq!(root_data(&tree), Raw(Str("foo".to_string())));
-    }
-
-    #[test]
-    fn test_decode_bigendianunicode() {
-        let mut tree = build_powershell_tree(
-            "[System.Text.Encoding]::BigEndianUnicode.getstring(@(0, 102, 0, 111, 0, 111))",
-        )
-        .unwrap();
-        tree.apply_mut(&mut (
-            Forward::default(),
-            ParseType::default(),
-            EncodingType::default(),
-            ParseInt::default(),
-            ParseArrayLiteral::default(),
-            ComputeArrayExpr::default(),
-            EncodingGetString::default(),
-        ))
-        .unwrap();
-
-        assert_eq!(root_data(&tree), Raw(Str("foo".to_string())));
-    }
-
-    #[test]
-    fn test_decode_utf32() {
-        let mut tree = build_powershell_tree(
-            "[System.Text.Encoding]::utf32.getstring(@(102, 0, 0, 0, 111, 0, 0, 0, 111, 0, 0, 0))",
-        )
-        .unwrap();
-        tree.apply_mut(&mut (
-            Forward::default(),
-            ParseType::default(),
-            EncodingType::default(),
-            ParseInt::default(),
-            ParseArrayLiteral::default(),
-            ComputeArrayExpr::default(),
-            EncodingGetString::default(),
-        ))
-        .unwrap();
-
-        assert_eq!(root_data(&tree), Raw(Str("foo".to_string())));
-    }
-
-    #[test]
-    fn test_decode_latin1() {
-        let mut tree =
-            build_powershell_tree("[System.Text.Encoding]::Latin1.getstring(@(102, 111, 111))")
-                .unwrap();
-        tree.apply_mut(&mut (
-            Forward::default(),
-            ParseType::default(),
-            EncodingType::default(),
-            ParseInt::default(),
-            ParseArrayLiteral::default(),
-            ComputeArrayExpr::default(),
-            EncodingGetString::default(),
-        ))
-        .unwrap();
-
-        assert_eq!(root_data(&tree), Raw(Str("foo".to_string())));
-    }
-
-    #[test]
-    fn test_decode_ascii_replaces_high_bytes() {
-        let mut tree =
-            build_powershell_tree("[System.Text.Encoding]::ascii.getstring(@(102, 200, 111))")
-                .unwrap();
-        tree.apply_mut(&mut (
-            Forward::default(),
-            ParseType::default(),
-            EncodingType::default(),
-            ParseInt::default(),
-            ParseArrayLiteral::default(),
-            ComputeArrayExpr::default(),
-            EncodingGetString::default(),
-        ))
-        .unwrap();
-
-        assert_eq!(root_data(&tree), Raw(Str("f?o".to_string())));
-    }
-
-    #[test]
-    fn test_decode_with_invoke() {
-        let mut tree = build_powershell_tree(
-            "[System.Text.Encoding]::'unicode'.'getstring'.invoke(@(102, 0, 111, 0, 111, 0))",
-        )
-        .unwrap();
-        tree.apply_mut(&mut (
-            Forward::default(),
-            ParseType::default(),
-            EncodingType::default(),
-            ParseInt::default(),
-            ParseString::default(),
-            ParseArrayLiteral::default(),
-            ComputeArrayExpr::default(),
-            EncodingGetString::default(),
-        ))
-        .unwrap();
-
-        assert_eq!(root_data(&tree), Raw(Str("foo".to_string())));
-    }
-
-    #[test]
-    fn test_getencoding_by_name() {
-        let mut tree = build_powershell_tree(
-            "[System.Text.Encoding]::GetEncoding('utf-8').getstring(@(102, 111, 111))",
-        )
-        .unwrap();
-        tree.apply_mut(&mut (
-            Forward::default(),
-            ParseType::default(),
-            ParseString::default(),
-            EncodingType::default(),
-            ParseInt::default(),
-            ParseArrayLiteral::default(),
-            ComputeArrayExpr::default(),
-            EncodingGetString::default(),
-        ))
-        .unwrap();
-
-        assert_eq!(root_data(&tree), Raw(Str("foo".to_string())));
-    }
-
-    #[test]
-    fn test_getencoding_by_codepage() {
-        let mut tree = build_powershell_tree(
-            "[System.Text.Encoding]::GetEncoding(28591).getstring(@(102, 111, 111))",
-        )
-        .unwrap();
-        tree.apply_mut(&mut (
-            Forward::default(),
-            ParseType::default(),
-            EncodingType::default(),
-            ParseInt::default(),
-            ParseArrayLiteral::default(),
-            ComputeArrayExpr::default(),
-            EncodingGetString::default(),
-        ))
-        .unwrap();
-
-        assert_eq!(root_data(&tree), Raw(Str("foo".to_string())));
-    }
-
-    #[test]
-    fn test_new_constructor() {
-        let mut tree =
-            build_powershell_tree("[System.Text.UTF8Encoding]::new().getstring(@(102, 111, 111))")
-                .unwrap();
-        tree.apply_mut(&mut (
-            Forward::default(),
-            ParseType::default(),
-            EncodingType::default(),
-            ParseInt::default(),
-            ParseArrayLiteral::default(),
-            ComputeArrayExpr::default(),
-            EncodingGetString::default(),
-        ))
-        .unwrap();
-
-        assert_eq!(root_data(&tree), Raw(Str("foo".to_string())));
-    }
-
-    #[test]
-    fn test_new_object_constructor() {
-        let mut tree = build_powershell_tree(
-            "(New-Object System.Text.UTF8Encoding).getstring(@(102, 111, 111))",
-        )
-        .unwrap();
-        tree.apply_mut(&mut (
-            Forward::default(),
-            ParseType::default(),
-            EncodingType::default(),
-            ParseInt::default(),
-            ParseArrayLiteral::default(),
-            ComputeArrayExpr::default(),
-            EncodingGetString::default(),
-        ))
-        .unwrap();
-
-        assert_eq!(root_data(&tree), Raw(Str("foo".to_string())));
-    }
-
-    #[test]
-    fn test_encode_utf8() {
-        let mut tree =
-            build_powershell_tree("[System.Text.Encoding]::UTF8.getbytes('foo')").unwrap();
-        tree.apply_mut(&mut (
-            Forward::default(),
-            ParseType::default(),
-            EncodingType::default(),
-            ParseString::default(),
-            EncodingGetBytes::default(),
-        ))
-        .unwrap();
-
-        assert_eq!(root_data(&tree), Bytes(vec![102, 111, 111]));
-    }
-
-    #[test]
-    fn test_encode_unicode() {
-        let mut tree =
-            build_powershell_tree("[System.Text.Encoding]::Unicode.getbytes('foo')").unwrap();
-        tree.apply_mut(&mut (
-            Forward::default(),
-            ParseType::default(),
-            EncodingType::default(),
-            ParseString::default(),
-            EncodingGetBytes::default(),
-        ))
-        .unwrap();
-
-        assert_eq!(root_data(&tree), Bytes(vec![102, 0, 111, 0, 111, 0]));
-    }
-
-    #[test]
-    fn test_encode_ascii_replaces_out_of_range() {
-        let source = format!("[System.Text.Encoding]::ASCII.getbytes('{}')", '\u{1234}');
-        let mut tree = build_powershell_tree(&source).unwrap();
-        tree.apply_mut(&mut (
-            Forward::default(),
-            ParseType::default(),
-            EncodingType::default(),
-            ParseString::default(),
-            EncodingGetBytes::default(),
-        ))
-        .unwrap();
-
-        assert_eq!(root_data(&tree), Bytes(vec![63]));
-    }
-
-    #[test]
-    fn test_decode_default_pure_ascii() {
-        let mut tree =
-            build_powershell_tree("[System.Text.Encoding]::Default.getstring(@(102, 111, 111))")
-                .unwrap();
-        tree.apply_mut(&mut (
-            Forward::default(),
-            ParseType::default(),
-            EncodingType::default(),
-            ParseInt::default(),
-            ParseArrayLiteral::default(),
-            ComputeArrayExpr::default(),
-            EncodingGetString::default(),
-        ))
-        .unwrap();
-
-        assert_eq!(root_data(&tree), Raw(Str("foo".to_string())));
-    }
-
-    #[test]
-    fn test_decode_default_non_ascii_is_not_resolved() {
-        let mut tree =
-            build_powershell_tree("[System.Text.Encoding]::Default.getstring(@(102, 200, 111))")
-                .unwrap();
-        tree.apply_mut(&mut (
-            Forward::default(),
-            ParseType::default(),
-            EncodingType::default(),
-            ParseInt::default(),
-            ParseArrayLiteral::default(),
-            ComputeArrayExpr::default(),
-            EncodingGetString::default(),
-        ))
-        .unwrap();
-
-        assert_eq!(
-            tree.root()
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .data(),
-            None
-        );
-    }
-}
