+++ core/src/ps/foreach.rs
@@ -245,215 +245,3 @@ impl<'a> RuleMut<'a> for ForEach {
         Ok(())
     }
 }
-
-#[cfg(test)]
-mod test {
-    use crate::ps::Powershell::Array;
-    use crate::ps::Value::{Num, Str};
-    use crate::ps::array::ParseArrayLiteral;
-    use crate::ps::build_powershell_tree;
-    use crate::ps::cast::Cast;
-    use crate::ps::foreach::{ForEach, PSItemInferrator};
-    use crate::ps::forward::Forward;
-    use crate::ps::integer::ParseInt;
-    use crate::ps::string::ParseString;
-    use crate::ps::typing::ParseType;
-
-    #[test]
-    fn test_foreach_transparent() {
-        let mut tree = build_powershell_tree("(1,2,3) | % {$_}").unwrap();
-        tree.apply_mut(&mut (
-            ParseInt::default(),
-            Forward::default(),
-            ParseArrayLiteral::default(),
-            PSItemInferrator::default(),
-            ForEach::default(),
-        ))
-        .unwrap();
-
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .data()
-                .expect("Inferred type"),
-            Array(vec![Num(1), Num(2), Num(3)])
-        );
-    }
-
-    #[test]
-    fn test_foreach_transparent_with_mixed_array() {
-        let mut tree = build_powershell_tree("(\"a\",2,3) | % {$_}").unwrap();
-        tree.apply_mut(&mut (
-            ParseInt::default(),
-            Forward::default(),
-            ParseString::default(),
-            ParseArrayLiteral::default(),
-            PSItemInferrator::default(),
-            ForEach::default(),
-        ))
-        .unwrap();
-
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .data()
-                .expect("Inferred type"),
-            Array(vec![Str("a".to_string()), Num(2), Num(3)])
-        );
-    }
-
-    #[test]
-    fn test_foreach_transparent_with_one_element() {
-        let mut tree = build_powershell_tree("(1) | % {$_}").unwrap();
-        tree.apply_mut(&mut (
-            ParseInt::default(),
-            Forward::default(),
-            ParseString::default(),
-            ParseArrayLiteral::default(),
-            PSItemInferrator::default(),
-            ForEach::default(),
-        ))
-        .unwrap();
-
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .data()
-                .expect("Inferred type"),
-            Array(vec![Num(1)])
-        );
-    }
-
-    #[test]
-    fn test_foreach_cast_with_one_element() {
-        let mut tree = build_powershell_tree("(0x61) | % {[char]$_}").unwrap();
-        tree.apply_mut(&mut (
-            ParseInt::default(),
-            Forward::default(),
-            ParseString::default(),
-            ParseArrayLiteral::default(),
-            PSItemInferrator::default(),
-            ForEach::default(),
-            Cast::default(),
-            ParseType::default(),
-        ))
-        .unwrap();
-
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .data()
-                .expect("Inferred type"),
-            Array(vec![Str("a".to_string())])
-        );
-    }
-
-    #[test]
-    fn test_foreach_cast_with_array() {
-        let mut tree = build_powershell_tree("(0x61, 0x62) | % {[char]$_}").unwrap();
-        tree.apply_mut(&mut (
-            ParseInt::default(),
-            Forward::default(),
-            ParseString::default(),
-            ParseArrayLiteral::default(),
-            PSItemInferrator::default(),
-            ForEach::default(),
-            Cast::default(),
-            ParseType::default(),
-        ))
-        .unwrap();
-
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .data()
-                .expect("Inferred type"),
-            Array(vec![Str("a".to_string()), Str("b".to_string())])
-        );
-    }
-
-    #[test]
-    fn test_foreach_cast_with_array_and_static_result() {
-        let mut tree = build_powershell_tree("(0x61, 0x62) | % {'z'; [char]$_}").unwrap();
-        tree.apply_mut(&mut (
-            ParseInt::default(),
-            Forward::default(),
-            ParseString::default(),
-            ParseArrayLiteral::default(),
-            PSItemInferrator::default(),
-            ForEach::default(),
-            Cast::default(),
-            ParseType::default(),
-        ))
-        .unwrap();
-
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .data()
-                .expect("Inferred type"),
-            Array(vec![
-                Str("z".to_string()),
-                Str("a".to_string()),
-                Str("z".to_string()),
-                Str("b".to_string())
-            ])
-        );
-    }
-
-    #[test]
-    fn test_foreach_case_insensitive_transparent() {
-        let mut tree = build_powershell_tree("(1,2,3) | fOrEacH-ObjECT {$_}").unwrap();
-        tree.apply_mut(&mut (
-            ParseInt::default(),
-            Forward::default(),
-            ParseArrayLiteral::default(),
-            PSItemInferrator::default(),
-            ForEach::default(),
-        ))
-        .unwrap();
-
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .data()
-                .expect("Inferred type"),
-            Array(vec![Num(1), Num(2), Num(3)])
-        );
-    }
-}
