+++ core/src/ps/integer.rs
@@ -217,143 +217,3 @@ impl<'a> RuleMut<'a> for MultInt {
         Ok(())
     }
 }
-
-#[cfg(test)]
-mod test {
-    use super::*;
-    use crate::ps::build_powershell_tree;
-    use crate::ps::forward::Forward;
-
-    #[test]
-    fn test_add_two_elements() {
-        let mut tree = build_powershell_tree("4 + 5").unwrap();
-        tree.apply_mut(&mut (ParseInt::default(), Forward::default(), AddInt::default()))
-            .unwrap();
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .data()
-                .expect("Inferred type"),
-            Raw(Num(9))
-        );
-    }
-
-    #[test]
-    fn test_add_three_elements() {
-        let mut tree = build_powershell_tree("4 + 5 + 9").unwrap();
-        tree.apply_mut(&mut (ParseInt::default(), Forward::default(), AddInt::default()))
-            .unwrap();
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .data()
-                .expect("Inferred type"),
-            Raw(Num(18))
-        );
-    }
-
-    #[test]
-    fn test_minus_two_elements() {
-        let mut tree = build_powershell_tree("4 - 5").unwrap();
-        tree.apply_mut(&mut (ParseInt::default(), Forward::default(), AddInt::default()))
-            .unwrap();
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .data()
-                .expect("Inferred type"),
-            Raw(Num(-1))
-        );
-    }
-
-    #[test]
-    fn test_minus_two_elements_with_unary_operators() {
-        let mut tree = build_powershell_tree("4 + -5").unwrap();
-        tree.apply_mut(&mut (ParseInt::default(), Forward::default(), AddInt::default()))
-            .unwrap();
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .data()
-                .expect("Inferred type"),
-            Raw(Num(-1))
-        );
-    }
-
-    #[test]
-    fn test_minus_two_elements_with_two_unary_operators() {
-        let mut tree = build_powershell_tree("-4 - 5").unwrap();
-        tree.apply_mut(&mut (ParseInt::default(), Forward::default(), AddInt::default()))
-            .unwrap();
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .data()
-                .expect("Inferred type"),
-            Raw(Num(-9))
-        );
-    }
-
-    #[test]
-    fn test_mul_two_elements() {
-        let mut tree = build_powershell_tree("4 * 5").unwrap();
-        tree.apply_mut(&mut (ParseInt::default(), Forward::default(), MultInt::default()))
-            .unwrap();
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .data()
-                .expect("Inferred type"),
-            Raw(Num(20))
-        );
-    }
-
-    #[test]
-    fn test_mul_three_elements() {
-        let mut tree = build_powershell_tree("4 * 5 * 10").unwrap();
-        tree.apply_mut(&mut (ParseInt::default(), Forward::default(), MultInt::default()))
-            .unwrap();
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .data()
-                .expect("Inferred type"),
-            Raw(Num(200))
-        );
-    }
-}
