+++ core/src/ps/linter.rs
@@ -570,42 +570,3 @@ impl<'a> Rule<'a> for RemoveUnusedVar {
         Ok(())
     }
 }
-
-#[cfg(test)]
-mod test {
-    use crate::ps::build_powershell_tree;
-    use crate::ps::forward::Forward;
-    use crate::ps::linter::Linter;
-    use crate::ps::string::ParseString;
-
-    fn assert_roundtrips(source: &str, expected_output: &str) {
-        let mut tree = build_powershell_tree(source).unwrap();
-        tree.apply_mut(&mut (ParseString::default(), Forward::default()))
-            .unwrap();
-
-        let mut ps_litter_view = Linter::default();
-        tree.apply(&mut ps_litter_view).unwrap();
-
-        assert_eq!(ps_litter_view.output, expected_output);
-    }
-
-    #[test]
-    fn test_escape_newline_on_output() {
-        assert_roundtrips("\"a`nb\"", "\"a`nb\"");
-    }
-
-    #[test]
-    fn test_escape_dollar_on_output() {
-        assert_roundtrips("\"a`$b\"", "\"a`$b\"");
-    }
-
-    #[test]
-    fn test_escape_backtick_on_output() {
-        assert_roundtrips("\"a``b\"", "\"a``b\"");
-    }
-
-    #[test]
-    fn test_escape_control_chars_on_output() {
-        assert_roundtrips("\"a`tb`0c\"", "\"a`tb`0c\"");
-    }
-}
