+++ core/src/ps/loops.rs
@@ -294,132 +294,3 @@ impl<'a> RuleMut<'a> for ForStatementFlowControl {
         Ok(())
     }
 }
-
-#[cfg(test)]
-mod test {
-    use crate::ps::LoopStatus::{Dead, OneTurn};
-    use crate::ps::Powershell::{Loop, Raw};
-    use crate::ps::Value::Num;
-    use crate::ps::bool::Comparison;
-    use crate::ps::build_powershell_tree;
-    use crate::ps::forward::Forward;
-    use crate::ps::integer::ParseInt;
-    use crate::ps::loops::{ForStatementCondition, ForStatementFlowControl};
-    use crate::ps::var::Var;
-
-    #[test]
-    fn test_dead_for_statement() {
-        let mut tree = build_powershell_tree("for ($i = 0; $i -gt 1; $i++) {}").unwrap();
-        tree.apply_mut(&mut (
-            Forward::default(),
-            ParseInt::default(),
-            Comparison::default(),
-            Var::default(),
-            ForStatementCondition::default(),
-        ))
-        .unwrap();
-
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .data()
-                .expect("Inferred type"),
-            Loop(Dead)
-        );
-    }
-
-    #[test]
-    fn test_one_turn_for_statement() {
-        let mut tree =
-            build_powershell_tree("for ($i = 0; $i -lt 1000; $i++) {$i; break; $i = $i - 1}")
-                .unwrap();
-        tree.apply_mut(&mut (
-            ParseInt::default(),
-            Forward::default(),
-            ForStatementFlowControl::default(),
-        ))
-        .unwrap();
-
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .data()
-                .expect("Inferred type"),
-            Loop(OneTurn)
-        );
-
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(8)
-                .unwrap()
-                .child(1)
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .smallest_child()
-                .data()
-                .expect("Inferred type"),
-            Raw(Num(0))
-        );
-    }
-
-    #[test]
-    fn test_unpredictable_for_statement() {
-        let mut tree = build_powershell_tree("for ($i = 0; $i -lt 10; $i++) {$i}").unwrap();
-        tree.apply_mut(&mut (
-            Forward::default(),
-            ParseInt::default(),
-            Var::default(),
-            ForStatementCondition::default(),
-        ))
-        .unwrap();
-
-        // The loop should not be infered
-        assert!(
-            tree.root()
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(3)
-                .unwrap()
-                .data()
-                .is_none()
-        );
-
-        // The statement should not be infered
-        assert!(
-            tree.root()
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(8)
-                .unwrap()
-                .child(1)
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .data()
-                .is_none()
-        );
-    }
-}
