+++ core/src/ps/method.rs
@@ -225,143 +225,3 @@ impl<'a> RuleMut<'a> for DecodeBase64 {
         Ok(())
     }
 }
-
-#[cfg(test)]
-mod test {
-    use crate::ps::Powershell;
-    use crate::ps::Powershell::Raw;
-    use crate::ps::Value::Num;
-    use crate::ps::array::{ComputeArrayExpr, ParseArrayLiteral};
-    use crate::ps::build_powershell_tree;
-    use crate::ps::forward::Forward;
-    use crate::ps::integer::ParseInt;
-    use crate::ps::method::{DecodeBase64, Length};
-    use crate::ps::string::ParseString;
-    use crate::ps::typing::ParseType;
-
-    #[test]
-    fn test_array_length() {
-        let mut tree = build_powershell_tree("@(1,2,3).length").unwrap();
-        tree.apply_mut(&mut (
-            ParseInt::default(),
-            Forward::default(),
-            ComputeArrayExpr::default(),
-            ParseArrayLiteral::default(),
-            Length::default(),
-        ))
-        .unwrap();
-
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .data()
-                .expect("Inferred type"),
-            Raw(Num(3))
-        );
-    }
-
-    #[test]
-    fn test_str_length() {
-        let mut tree = build_powershell_tree("'foo'.length").unwrap();
-        tree.apply_mut(&mut (
-            ParseString::default(),
-            Forward::default(),
-            ComputeArrayExpr::default(),
-            ParseArrayLiteral::default(),
-            Length::default(),
-        ))
-        .unwrap();
-
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .data()
-                .expect("Inferred type"),
-            Raw(Num(3))
-        );
-    }
-
-    #[test]
-    fn test_decode_base64() {
-        let mut tree = build_powershell_tree("[System.Convert]::FromBase64String('Zm9v')").unwrap();
-        tree.apply_mut(&mut (
-            ParseString::default(),
-            Forward::default(),
-            DecodeBase64::default(),
-            ParseType::default(),
-        ))
-        .unwrap();
-
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .data()
-                .expect("Inferred type"),
-            Powershell::Bytes(vec![102, 111, 111])
-        );
-    }
-
-    #[test]
-    fn test_error_decode_base64() {
-        let mut tree =
-            build_powershell_tree("[System.Convert]::FromBase64String('AAAAAAAAAA')").unwrap();
-        tree.apply_mut(&mut (
-            ParseString::default(),
-            Forward::default(),
-            DecodeBase64::default(),
-            ParseType::default(),
-        ))
-        .unwrap();
-
-        assert_eq!(
-            tree.root()
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .data(),
-            None
-        );
-    }
-
-    #[test]
-    fn test_error_decode_base64_with_invoke() {
-        let mut tree =
-            build_powershell_tree("[System.Convert]::'FromBase64String'.invoke('AAAAAAAAAA')")
-                .unwrap();
-        tree.apply_mut(&mut (
-            ParseString::default(),
-            Forward::default(),
-            DecodeBase64::default(),
-            ParseType::default(),
-        ))
-        .unwrap();
-
-        assert_eq!(
-            tree.root()
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .child(0)
-                .unwrap()
-                .data(),
-            None
-        );
-    }
-}
