+++ core/src/ps/var.rs
@@ -765,1519 +765,3 @@ impl<'a> Rule<'a> for UnusedVar {
         Ok(())
     }
 }
-
-#[cfg(test)]
-mod test {
-    use super::*;
-    use crate::ps::access::AccessHashMap;
-    use crate::ps::array::ParseArrayLiteral;
-    use crate::ps::bool::ParseBool;
-    use crate::ps::build_powershell_tree;
-    use crate::ps::forward::Forward;
-    use crate::ps::hash::ParseHash;
-    use crate::ps::integer::{AddInt, ParseInt};
-    use crate::ps::strategy::PowershellStrategy;
-    use crate::ps::string::ParseString;
-
-    #[test]
-    fn test_static_replacement() {
-        let mut tree = build_powershell_tree("$foo = 4\nWrite-Debug $foo").unwrap();
-
-        tree.apply_mut_with_strategy(
-            &mut (ParseInt::default(), Forward::default(), Var::default()),
-            PowershellStrategy::default(),
-        )
-        .unwrap();
-
-        // We are waiting for
-        // Write-Debug 4
-        // (program
-        //  (statement_list inferred_type: None)
-        //   (pipeline inferred_type: None
-        //    (pipeline_chain inferred_type: None
-        //     (command inferred_type: None
-        //      (command_name inferred_type: None)
-        //      (command_elements inferred_type: None)
-        //       (command_param_sep)
-        //       (variable inferred_type: Some(Number(4)))))))))
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap() // program
-                .child(0)
-                .unwrap() // statement_list
-                .child(1)
-                .unwrap() // pipeline
-                .child(0)
-                .unwrap() // pipeline_chain
-                .child(0)
-                .unwrap() //command
-                .child(1)
-                .unwrap() // command_elements
-                .child(1)
-                .unwrap() // variable
-                .data()
-                .expect("Expecting inferred type"),
-            Raw(Num(4))
-        );
-    }
-
-    #[test]
-    fn test_unfollow_var_use_unknow_var() {
-        let mut tree = build_powershell_tree("$foo = $toto\nWrite-Debug $foo").unwrap();
-
-        tree.apply_mut_with_strategy(
-            &mut (ParseInt::default(), Forward::default(), Var::default()),
-            PowershellStrategy::default(),
-        )
-        .unwrap();
-
-        // We are waiting for
-        // Write-Debug 4
-        // (program
-        //  (statement_list inferred_type: None)
-        //   (pipeline inferred_type: None
-        //    (pipeline_chain inferred_type: None
-        //     (command inferred_type: None
-        //      (command_name inferred_type: None)
-        //      (command_elements inferred_type: None)
-        //       (variable inferred_type: Some(Number(4)))))))))
-        assert_eq!(
-            tree.root()
-                .unwrap()
-                .child(0)
-                .unwrap() // statement_list
-                .child(1)
-                .unwrap() // pipeline
-                .child(0)
-                .unwrap() // pipeline_chain
-                .child(0)
-                .unwrap() //command
-                .child(1)
-                .unwrap() // command_elements
-                .child(0)
-                .unwrap() // variable
-                .data(),
-            None
-        );
-    }
-
-    #[test]
-    fn test_static_var_shell_id() {
-        let mut tree = build_powershell_tree("$shellid").unwrap();
-
-        tree.apply_mut_with_strategy(
-            &mut (
-                ParseInt::default(),
-                Forward::default(),
-                StaticVar::default(),
-            ),
-            PowershellStrategy::default(),
-        )
-        .unwrap();
-
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap() // program
-                .child(0)
-                .unwrap() // statement_list
-                .child(0)
-                .unwrap() // pipeline
-                .data()
-                .expect("Expecting inferred type"),
-            Raw(Str("Microsoft.Powershell".to_string()))
-        );
-    }
-
-    #[test]
-    fn test_unfollow_var_use_in_if_statement() {
-        let mut tree =
-            build_powershell_tree("$foo = 0\nif(unknown) { $foo = 5 }\n White-Debug $foo").unwrap();
-
-        tree.apply_mut_with_strategy(
-            &mut (ParseInt::default(), Forward::default(), Var::default()),
-            PowershellStrategy::default(),
-        )
-        .unwrap();
-
-        // We are waiting for
-        // Write-Debug $foo
-        // (program
-        //  (statement_list inferred_type: None)
-        //   (pipeline inferred_type: None
-        //    (pipeline_chain inferred_type: None
-        //     (command inferred_type: None
-        //      (command_name inferred_type: None)
-        //      (command_elements inferred_type: None)
-        //       (command_argument_sep)
-        //       (variable inferred_type: None)))))))
-        assert_eq!(
-            tree.root()
-                .unwrap()
-                .child(0)
-                .unwrap() // statement_list
-                .child(2)
-                .unwrap() // pipeline
-                .child(0)
-                .unwrap() // pipeline_chain
-                .child(0)
-                .unwrap() //command
-                .child(1)
-                .unwrap() // command_elements
-                .child(1)
-                .unwrap() // variable
-                .data(),
-            None
-        );
-    }
-
-    #[test]
-    fn test_infer_var_use_in_if_statement_predictable() {
-        let mut tree =
-            build_powershell_tree("$foo = 0\nif($true) { $foo = 5 }\nWhite-Debug $foo").unwrap();
-
-        tree.apply_mut_with_strategy(
-            &mut (
-                ParseInt::default(),
-                Forward::default(),
-                Var::default(),
-                ParseBool::default(),
-            ),
-            PowershellStrategy::default(),
-        )
-        .unwrap();
-
-        // We are waiting for
-        // Write-Debug 5
-        // (program
-        //  (statement_list inferred_type: None)
-        //   (pipeline inferred_type: None
-        //    (pipeline_chain inferred_type: None
-        //     (command inferred_type: None
-        //      (command_name inferred_type: None)
-        //      (command_elements inferred_type: None)
-        //       (command_argument_sep)
-        //       (variable inferred_type: Some(Num(5)))))))
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap() // statement_list
-                .child(2)
-                .unwrap() // pipeline
-                .child(0)
-                .unwrap() // pipeline_chain
-                .child(0)
-                .unwrap() //command
-                .child(1)
-                .unwrap() // command_elements
-                .child(1)
-                .unwrap() // variable
-                .data()
-                .expect("Expecting inferred type"),
-            Raw(Num(5))
-        );
-    }
-
-    #[test]
-    fn test_infer_var_use_in_if_statement_predictable_false() {
-        let mut tree =
-            build_powershell_tree("$foo = 0\nif($false) { $foo = 5 }\nWhite-Debug $foo").unwrap();
-
-        tree.apply_mut_with_strategy(
-            &mut (
-                ParseInt::default(),
-                Forward::default(),
-                Var::default(),
-                ParseBool::default(),
-            ),
-            PowershellStrategy::default(),
-        )
-        .unwrap();
-
-        // We are waiting for
-        // Write-Debug 0
-        // (program
-        //  (statement_list inferred_type: None)
-        //   (pipeline inferred_type: None
-        //    (pipeline_chain inferred_type: None
-        //     (command inferred_type: None
-        //      (command_name inferred_type: None)
-        //      (command_elements inferred_type: None)
-        //       (command_argument_sep)
-        //       (variable inferred_type: Some(Num(0)))))))
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap() // statement_list
-                .child(2)
-                .unwrap() // pipeline
-                .child(0)
-                .unwrap() // pipeline_chain
-                .child(0)
-                .unwrap() //command
-                .child(1)
-                .unwrap() // command_elements
-                .child(1)
-                .unwrap() // variable
-                .data()
-                .expect("Expecting inferred type"),
-            Raw(Num(0))
-        );
-    }
-
-    #[test]
-    fn test_infer_var_use_in_if_else_statement_predictable() {
-        let mut tree = build_powershell_tree(
-            "$foo = 0\nif($false) { $foo = 5 }else { $foo = 8 }\nWhite-Debug $foo",
-        )
-        .unwrap();
-
-        tree.apply_mut_with_strategy(
-            &mut (
-                ParseInt::default(),
-                Forward::default(),
-                Var::default(),
-                ParseBool::default(),
-            ),
-            PowershellStrategy::default(),
-        )
-        .unwrap();
-
-        // We are waiting for
-        // Write-Debug 8
-        // (program
-        //  (statement_list inferred_type: None)
-        //   (pipeline inferred_type: None
-        //    (pipeline_chain inferred_type: None
-        //     (command inferred_type: None
-        //      (command_name inferred_type: None)
-        //      (command_elements inferred_type: None)
-        //       (command_argument_sep)
-        //       (variable inferred_type: Some(Num(8)))))))
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap() // statement_list
-                .child(2)
-                .unwrap() // pipeline
-                .child(0)
-                .unwrap() // pipeline_chain
-                .child(0)
-                .unwrap() //command
-                .child(1)
-                .unwrap() // command_elements
-                .child(1)
-                .unwrap() // variable
-                .data()
-                .expect("Expecting inferred type"),
-            Raw(Num(8))
-        );
-    }
-
-    #[test]
-    fn test_infer_var_use_in_if_elseif_else_statement_predictable() {
-        let mut tree = build_powershell_tree("$foo = 0\nif($false) { $foo = 5 }elseif($true) { $foo = 6 } else {$foo = 7}\nWhite-Debug $foo").unwrap();
-
-        tree.apply_mut_with_strategy(
-            &mut (
-                ParseInt::default(),
-                Forward::default(),
-                Var::default(),
-                ParseBool::default(),
-            ),
-            PowershellStrategy::default(),
-        )
-        .unwrap();
-
-        // We are waiting for
-        // Write-Debug 6
-        // (program
-        //  (statement_list inferred_type: None)
-        //   (pipeline inferred_type: None
-        //    (pipeline_chain inferred_type: None
-        //     (command inferred_type: None
-        //      (command_name inferred_type: None)
-        //      (command_elements inferred_type: None)
-        //       (command_argument_sep)
-        //       (variable inferred_type: Some(Num(6)))))))
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap() // statement_list
-                .child(2)
-                .unwrap() // pipeline
-                .child(0)
-                .unwrap() // pipeline_chain
-                .child(0)
-                .unwrap() //command
-                .child(1)
-                .unwrap() // command_elements
-                .child(1)
-                .unwrap() // variable
-                .data()
-                .expect("Expecting inferred type"),
-            Raw(Num(6))
-        );
-    }
-
-    #[test]
-    fn test_infer_var_use_in_if_elseif_else_statement_unpredictable() {
-        let mut tree = build_powershell_tree("$foo = 0\nif($false) { $foo = 5 }elseif(unknown) { $foo = 6 } else {$foo = 7}\nWhite-Debug $foo").unwrap();
-
-        tree.apply_mut_with_strategy(
-            &mut (
-                ParseInt::default(),
-                Forward::default(),
-                Var::default(),
-                ParseBool::default(),
-            ),
-            PowershellStrategy::default(),
-        )
-        .unwrap();
-
-        // We are waiting for
-        // Write-Debug 5
-        // (program
-        //  (statement_list inferred_type: None)
-        //   (pipeline inferred_type: None
-        //    (pipeline_chain inferred_type: None
-        //     (command inferred_type: None
-        //      (command_name inferred_type: None)
-        //      (command_elements inferred_type: None)
-        //       (command_argument_sep)
-        //       (variable inferred_type: Some(Num(5)))))))
-        assert_eq!(
-            tree.root()
-                .unwrap()
-                .child(0)
-                .unwrap() // statement_list
-                .child(2)
-                .unwrap() // pipeline
-                .child(0)
-                .unwrap() // pipeline_chain
-                .child(0)
-                .unwrap() //command
-                .child(1)
-                .unwrap() // command_elements
-                .child(1)
-                .unwrap() // variable
-                .data(),
-            None
-        );
-    }
-
-    #[test]
-    fn test_infer_var_use_in_if_elseif_else_statement_predictable_in_else() {
-        let mut tree = build_powershell_tree("$foo = 0\nif($false) { $foo = 5 }elseif($false) { $foo = 6 } else {$foo = 7}\nWhite-Debug $foo").unwrap();
-
-        tree.apply_mut_with_strategy(
-            &mut (
-                ParseInt::default(),
-                Forward::default(),
-                Var::default(),
-                ParseBool::default(),
-            ),
-            PowershellStrategy::default(),
-        )
-        .unwrap();
-
-        // We are waiting for
-        // Write-Debug 7
-        // (program
-        //  (statement_list inferred_type: None)
-        //   (pipeline inferred_type: None
-        //    (pipeline_chain inferred_type: None
-        //     (command inferred_type: None
-        //      (command_name inferred_type: None)
-        //      (command_elements inferred_type: None)
-        //       (command_argument_sep)
-        //       (variable inferred_type: Some(Num(7)))))))
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap() // statement_list
-                .child(2)
-                .unwrap() // pipeline
-                .child(0)
-                .unwrap() // pipeline_chain
-                .child(0)
-                .unwrap() //command
-                .child(1)
-                .unwrap() // command_elements
-                .child(1)
-                .unwrap() // variable
-                .data()
-                .expect("Expecting inferred type"),
-            Raw(Num(7))
-        );
-    }
-
-    #[test]
-    fn test_infer_var_use_in_while_statement_use_in_statement() {
-        // var is used in the loop statement -> not inferred in the condition and forget
-        let mut tree =
-            build_powershell_tree("$a = 1\nwhile($a -gt 0) { $a = $a + 1 }\nWhite-Debug $a")
-                .unwrap();
-
-        tree.apply_mut_with_strategy(
-            &mut (
-                ParseInt::default(),
-                Forward::default(),
-                Var::default(),
-                ParseBool::default(),
-            ),
-            PowershellStrategy::default(),
-        )
-        .unwrap();
-
-        // We are waiting for
-        // Write-Debug 5
-        // (program
-        //  (statement_list inferred_type: None)
-        //   (pipeline inferred_type: None
-        //    (pipeline_chain inferred_type: None
-        //     (command inferred_type: None
-        //      (command_name inferred_type: None)
-        //      (command_elements inferred_type: None)
-        //       (command_argument_sep)
-        //       (variable inferred_type: Some(Num(5)))))))
-        assert_eq!(
-            tree.root()
-                .unwrap()
-                .child(0)
-                .unwrap() // statement_list
-                .child(2)
-                .unwrap() // pipeline
-                .child(0)
-                .unwrap() // pipeline_chain
-                .child(0)
-                .unwrap() //command
-                .child(1)
-                .unwrap() // command_elements
-                .child(1)
-                .unwrap() // variable
-                .data(),
-            None
-        );
-    }
-
-    #[test]
-    fn test_infer_var_use_in_while_statement_not_use_in_statement() {
-        // var is used in the loop statement -> not inferred in the condition and forget
-        let mut tree =
-            build_powershell_tree("$a = 1\nwhile($a -gt 0) { $b = $a + 1 }\nWhite-Debug $a")
-                .unwrap();
-
-        tree.apply_mut_with_strategy(
-            &mut (
-                ParseInt::default(),
-                Forward::default(),
-                Var::default(),
-                ParseBool::default(),
-            ),
-            PowershellStrategy::default(),
-        )
-        .unwrap();
-
-        // We are waiting for
-        // Write-Debug 5
-        // (program
-        //  (statement_list inferred_type: None)
-        //   (pipeline inferred_type: None
-        //    (pipeline_chain inferred_type: None
-        //     (command inferred_type: None
-        //      (command_name inferred_type: None)
-        //      (command_elements inferred_type: None)
-        //       (command_argument_sep)
-        //       (variable inferred_type: Some(Num(5)))))))
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap() // statement_list
-                .child(2)
-                .unwrap() // pipeline
-                .child(0)
-                .unwrap() // pipeline_chain
-                .child(0)
-                .unwrap() //command
-                .child(1)
-                .unwrap() // command_elements
-                .child(1)
-                .unwrap() // variable
-                .data()
-                .expect("Expecting inferred type"),
-            Raw(Num(1))
-        );
-    }
-
-    #[test]
-    fn test_infer_var_use_in_function_statement() {
-        // infer global var in function statement
-        let mut tree = build_powershell_tree("$a = 1\nFunction invoke { $a }").unwrap();
-
-        tree.apply_mut_with_strategy(
-            &mut (
-                ParseInt::default(),
-                Forward::default(),
-                Var::default(),
-                ParseBool::default(),
-            ),
-            PowershellStrategy::default(),
-        )
-        .unwrap();
-
-        // We are waiting for
-        // Write-Debug 5
-        // (program
-        //  (statement_list inferred_type: None)
-        //   (assignment_expression inferred_type: None) ...
-        //   (function_statement inferred_type: None
-        //    (function inferred_type: None)
-        //    (function_name inferred_type: None)
-        //    ({ inferred_type: None)
-        //    (script_block inferred_type: None
-        //     (script_block_body inferred_type: None
-        //      (statement_list inferred_type: None
-        //       (pipeline inferred_type: Some(Raw(Num(1)))
-        //        (logical_expression inferred_type: Some(Raw(Num(1)))
-        //         (bitwise_expression inferred_type: None
-        //          (comparison_expression inferred_type: None
-        //           (additive_expression inferred_type: None
-        //            (multiplicative_expression inferred_type: None
-        //             (format_expression inferred_type: None
-        //              (range_expression inferred_type: None
-        //               (array_literal_expression inferred_type: None
-        //                (unary_expression inferred_type: None
-        //                 (variable inferred_type: None))))))))))))))
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap() // statement_list
-                .child(1)
-                .unwrap() // function_statement
-                .child(3)
-                .unwrap() //script_block
-                .child(0)
-                .unwrap() // script_block_body
-                .child(0)
-                .unwrap() // statement_list
-                .child(0)
-                .unwrap() // pipeline
-                .data()
-                .expect("Expecting inferred type"),
-            Raw(Num(1))
-        );
-    }
-
-    #[test]
-    fn test_wildcarded_variable() {
-        // infer global var in function statement
-        let mut tree = build_powershell_tree("sV my-var 1\n(varIable M*ar).vaLue").unwrap();
-
-        tree.apply_mut_with_strategy(
-            &mut (
-                ParseInt::default(),
-                Forward::default(),
-                Var::default(),
-                ParseHash::default(),
-                AccessHashMap::default(),
-            ),
-            PowershellStrategy::default(),
-        )
-        .unwrap();
-
-        // We are waiting for
-        // (program inferred_type: None
-        //  (statement_list inferred_type: None
-        //   (pipeline inferred_type: None
-        //    (command inferred_type: None
-        //     (command_name inferred_type: None)
-        //     (command_elements inferred_type: None
-        //      (command_argument_sep inferred_type: None
-        //       (  inferred_type: None))
-        //      (generic_token inferred_type: None)
-        //      (command_argument_sep inferred_type: None
-        //       (  inferred_type: None))
-        //      (array_literal_expression inferred_type: Some(Raw(Num(1)))
-        //       (unary_expression inferred_type: None
-        //        (integer_literal inferred_type: None
-        //         (decimal_integer_literal inferred_type: None)))))))
-        //   (pipeline inferred_type: Some(Raw(Num(1)))
-        //    (logical_expression inferred_type: Some(Raw(Num(1)))
-        //     (bitwise_expression inferred_type: None
-        //      (comparison_expression inferred_type: None
-        //       (additive_expression inferred_type: None
-        //        (multiplicative_expression inferred_type: None
-        //         (format_expression inferred_type: None
-        //          (range_expression inferred_type: None
-        //           (array_literal_expression inferred_type: None
-        //            (unary_expression inferred_type: None
-        //             (member_access inferred_type: None
-        //              (parenthesized_expression inferred_type: Some(HashMap({Str("name"): Str("my-var"), Str("value"): Num(1)}))
-        //               (( inferred_type: None)
-        //               (pipeline inferred_type: None
-        //                (command inferred_type: Some(HashMap({Str("name"): Str("my-var"), Str("value"): Num(1)}))
-        //                 (command_name inferred_type: None)
-        //                 (command_elements inferred_type: None
-        //                  (command_argument_sep inferred_type: None
-        //                   (  inferred_type: None))
-        //                  (generic_token inferred_type: None))))
-        //               () inferred_type: None))
-        //              (. inferred_type: None)
-        //              (member_name inferred_type: None
-        //               (simple_name inferred_type: None)))))))))))))))
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap() // statement_list
-                .child(1)
-                .unwrap() // pipeline
-                .data()
-                .expect("Expecting inferred type"),
-            Raw(Num(1))
-        );
-    }
-
-    #[test]
-    fn test_wildcarded_getvariable() {
-        // infer global var in function statement
-        let mut tree = build_powershell_tree("sV my-var 1\ngV M*ar -vaL").unwrap();
-
-        tree.apply_mut_with_strategy(
-            &mut (
-                ParseInt::default(),
-                Forward::default(),
-                Var::default(),
-                ParseHash::default(),
-                AccessHashMap::default(),
-            ),
-            PowershellStrategy::default(),
-        )
-        .unwrap();
-
-        // We are waiting for
-        // (program inferred_type: None
-        // (statement_list inferred_type: None
-        //  (pipeline inferred_type: None
-        //   (command inferred_type: None
-        //    (command_name inferred_type: None)
-        //    (command_elements inferred_type: None
-        //     (command_argument_sep inferred_type: None
-        //      (  inferred_type: None))
-        //     (generic_token inferred_type: None)
-        //     (command_argument_sep inferred_type: None
-        //      (  inferred_type: None))
-        //     (array_literal_expression inferred_type: Some(Raw(Num(1)))
-        //      (unary_expression inferred_type: None
-        //       (integer_literal inferred_type: None
-        //        (decimal_integer_literal inferred_type: None)))))))
-        //  (pipeline inferred_type: Some(Raw(Num(1)))
-        //   (command inferred_type: Some(Raw(Num(1)))
-        //    (command_name inferred_type: None)
-        //    (command_elements inferred_type: None
-        //     (command_argument_sep inferred_type: None
-        //      (  inferred_type: None))
-        //     (generic_token inferred_type: None)
-        //     (command_argument_sep inferred_type: None
-        //      (  inferred_type: None))
-        //     (command_parameter inferred_type: None))))))
-
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap() // statement_list
-                .child(1)
-                .unwrap() // pipeline
-                .data()
-                .expect("Expecting inferred type"),
-            Raw(Num(1))
-        );
-    }
-
-    #[test]
-    fn test_wildcarded_getsetitem() {
-        // infer global var in function statement
-        let mut tree =
-            build_powershell_tree("sv mYVAr 1\nsi variable:/M*ar 2\n(ls variable:*y*ar).value")
-                .unwrap();
-
-        tree.apply_mut_with_strategy(
-            &mut (
-                ParseInt::default(),
-                Forward::default(),
-                Var::default(),
-                ParseHash::default(),
-                AccessHashMap::default(),
-            ),
-            PowershellStrategy::default(),
-        )
-        .unwrap();
-
-        // We are waiting for
-        // (program inferred_type: None
-        // (statement_list inferred_type: None
-        //  (pipeline inferred_type: None
-        //   (command inferred_type: None
-        //    (command_name inferred_type: None)
-        //    (command_elements inferred_type: None
-        //     (command_argument_sep inferred_type: None
-        //      (  inferred_type: None))
-        //     (generic_token inferred_type: None)
-        //     (command_argument_sep inferred_type: None
-        //      (  inferred_type: None))
-        //     (array_literal_expression inferred_type: Some(Raw(Num(1)))
-        //      (unary_expression inferred_type: None
-        //       (integer_literal inferred_type: None
-        //        (decimal_integer_literal inferred_type: None)))))))
-        //  (pipeline inferred_type: None
-        //   (command inferred_type: None
-        //    (command_name inferred_type: None)
-        //    (command_elements inferred_type: None
-        //     (command_argument_sep inferred_type: None
-        //      (  inferred_type: None))
-        //     (generic_token inferred_type: None)
-        //     (command_argument_sep inferred_type: None
-        //      (  inferred_type: None))
-        //     (array_literal_expression inferred_type: Some(Raw(Num(2)))
-        //      (unary_expression inferred_type: None
-        //       (integer_literal inferred_type: None
-        //        (decimal_integer_literal inferred_type: None)))))))
-        //  (pipeline inferred_type: Some(Raw(Num(2)))
-        //   (logical_expression inferred_type: Some(Raw(Num(2)))
-        //    (bitwise_expression inferred_type: None
-        //     (comparison_expression inferred_type: None
-        //      (additive_expression inferred_type: None
-        //       (multiplicative_expression inferred_type: None
-        //        (format_expression inferred_type: None
-        //         (range_expression inferred_type: None
-        //          (array_literal_expression inferred_type: None
-        //           (unary_expression inferred_type: None
-        //            (member_access inferred_type: None
-        //             (parenthesized_expression inferred_type: Some(HashMap({Str("name"): Str("myvar"), Str("value"): Num(2)}))
-        //              (( inferred_type: None)
-        //              (pipeline inferred_type: None
-        //               (command inferred_type: Some(HashMap({Str("name"): Str("myvar"), Str("value"): Num(2)}))
-        //                (command_name inferred_type: None)
-        //                (command_elements inferred_type: None
-        //                 (command_argument_sep inferred_type: None
-        //                  (  inferred_type: None))
-        //                 (generic_token inferred_type: None))))
-        //              () inferred_type: None))
-        //             (. inferred_type: None)
-        //             (member_name inferred_type: None
-        //              (simple_name inferred_type: None)))))))))))))))
-
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap() // statement_list
-                .child(2)
-                .unwrap() // pipeline
-                .data()
-                .expect("Expecting inferred type"),
-            Raw(Num(2))
-        );
-    }
-
-    #[test]
-    fn test_add_assignment_operator_int_int() {
-        // infer global var in function statement
-        let mut tree = build_powershell_tree("$a=1;$a+=1;$a").unwrap();
-
-        tree.apply_mut_with_strategy(
-            &mut (ParseInt::default(), Forward::default(), Var::default()),
-            PowershellStrategy::default(),
-        )
-        .unwrap();
-
-        // We are waiting for
-        // (program inferred_type: None
-        //  (statement_list inferred_type: None
-        //   (pipeline inferred_type: None
-        //     ...)
-        //   (empty_statement inferred_type: None
-        //    ...)
-        //   (pipeline inferred_type: None
-        //    ...)
-        //   (empty_statement inferred_type: None
-        //    ...)
-        //   (pipeline inferred_type: Some(Raw(Num(2)))  <--- correct infered type
-        //    ...)
-
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap() // statement_list
-                .child(4)
-                .unwrap() // pipeline
-                .data()
-                .expect("Expecting inferred type"),
-            Raw(Num(2))
-        );
-    }
-    #[test]
-    fn test_add_assignment_operator_int_str() {
-        // infer global var in function statement
-        let mut tree = build_powershell_tree("$a=1;$a+=\"1\";$a").unwrap();
-
-        tree.apply_mut_with_strategy(
-            &mut (
-                ParseInt::default(),
-                ParseString::default(),
-                Forward::default(),
-                Var::default(),
-            ),
-            PowershellStrategy::default(),
-        )
-        .unwrap();
-
-        // We are waiting for
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap() // statement_list
-                .child(4)
-                .unwrap() // pipeline
-                .data()
-                .expect("Expecting inferred type"),
-            Raw(Num(2))
-        );
-    }
-    #[test]
-    fn test_add_assignment_operator_str_int() {
-        // infer global var in function statement
-        let mut tree = build_powershell_tree("$a=\"1\";$a+=1;$a").unwrap();
-
-        tree.apply_mut_with_strategy(
-            &mut (
-                ParseInt::default(),
-                ParseString::default(),
-                Forward::default(),
-                Var::default(),
-            ),
-            PowershellStrategy::default(),
-        )
-        .unwrap();
-
-        // We are waiting for
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap() // statement_list
-                .child(4)
-                .unwrap() // pipeline
-                .data()
-                .expect("Expecting inferred type"),
-            Raw(Str(String::from("11")))
-        );
-    }
-    #[test]
-    fn test_add_assignment_operator_str_str() {
-        // infer global var in function statement
-        let mut tree = build_powershell_tree("$a=\"1\";$a+=\"1\";$a").unwrap();
-
-        tree.apply_mut_with_strategy(
-            &mut (
-                ParseInt::default(),
-                ParseString::default(),
-                Forward::default(),
-                Var::default(),
-            ),
-            PowershellStrategy::default(),
-        )
-        .unwrap();
-
-        // We are waiting for
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap() // statement_list
-                .child(4)
-                .unwrap() // pipeline
-                .data()
-                .expect("Expecting inferred type"),
-            Raw(Str(String::from("11")))
-        );
-    }
-
-    #[test]
-    fn test_sub_assignment_operator_int_int() {
-        // infer global var in function statement
-        let mut tree = build_powershell_tree("$a=1;$a-=1;$a").unwrap();
-
-        tree.apply_mut_with_strategy(
-            &mut (ParseInt::default(), Forward::default(), Var::default()),
-            PowershellStrategy::default(),
-        )
-        .unwrap();
-
-        // We are waiting for
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap() // statement_list
-                .child(4)
-                .unwrap() // pipeline
-                .data()
-                .expect("Expecting inferred type"),
-            Raw(Num(0))
-        );
-    }
-
-    #[test]
-    fn test_sub_assignment_operator_int_str() {
-        // infer global var in function statement
-        let mut tree = build_powershell_tree("$a=1;$a-=\"1\";$a").unwrap();
-
-        tree.apply_mut_with_strategy(
-            &mut (
-                ParseInt::default(),
-                ParseString::default(),
-                Forward::default(),
-                Var::default(),
-            ),
-            PowershellStrategy::default(),
-        )
-        .unwrap();
-
-        // We are waiting for
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap() // statement_list
-                .child(4)
-                .unwrap() // pipeline
-                .data()
-                .expect("Expecting inferred type"),
-            Raw(Num(0))
-        );
-    }
-    #[test]
-    fn test_sub_assignment_operator_str_int() {
-        // infer global var in function statement
-        let mut tree = build_powershell_tree("$a=\"1\";$a-=1;$a").unwrap();
-
-        tree.apply_mut_with_strategy(
-            &mut (
-                ParseInt::default(),
-                ParseString::default(),
-                Forward::default(),
-                Var::default(),
-            ),
-            PowershellStrategy::default(),
-        )
-        .unwrap();
-
-        // We are waiting for
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap() // statement_list
-                .child(4)
-                .unwrap() // pipeline
-                .data()
-                .expect("Expecting inferred type"),
-            Raw(Num(0))
-        );
-    }
-    #[test]
-    fn test_sub_assignment_operator_str_str() {
-        // infer global var in function statement
-        let mut tree = build_powershell_tree("$a=\"1\";$a-=\"1\";$a").unwrap();
-
-        tree.apply_mut_with_strategy(
-            &mut (
-                ParseInt::default(),
-                ParseString::default(),
-                Forward::default(),
-                Var::default(),
-            ),
-            PowershellStrategy::default(),
-        )
-        .unwrap();
-
-        // We are waiting for
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap() // statement_list
-                .child(4)
-                .unwrap() // pipeline
-                .data()
-                .expect("Expecting inferred type"),
-            Raw(Num(0))
-        );
-    }
-
-    #[test]
-    fn test_mul_assignment_operator_int_int() {
-        // infer global var in function statement
-        let mut tree = build_powershell_tree("$a=2;$a*=3;$a").unwrap();
-
-        tree.apply_mut_with_strategy(
-            &mut (ParseInt::default(), Forward::default(), Var::default()),
-            PowershellStrategy::default(),
-        )
-        .unwrap();
-
-        // We are waiting for
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap() // statement_list
-                .child(4)
-                .unwrap() // pipeline
-                .data()
-                .expect("Expecting inferred type"),
-            Raw(Num(6))
-        );
-    }
-
-    #[test]
-    fn test_mul_assignment_operator_int_str() {
-        // infer global var in function statement
-        let mut tree = build_powershell_tree("$a=2;$a*=\"3\";$a").unwrap();
-
-        tree.apply_mut_with_strategy(
-            &mut (
-                ParseInt::default(),
-                ParseString::default(),
-                Forward::default(),
-                Var::default(),
-            ),
-            PowershellStrategy::default(),
-        )
-        .unwrap();
-
-        // We are waiting for
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap() // statement_list
-                .child(4)
-                .unwrap() // pipeline
-                .data()
-                .expect("Expecting inferred type"),
-            Raw(Num(6))
-        );
-    }
-    #[test]
-    fn test_mul_assignment_operator_str_int() {
-        // infer global var in function statement
-        let mut tree = build_powershell_tree("$a=\"12\";$a*=3;$a").unwrap();
-
-        tree.apply_mut_with_strategy(
-            &mut (
-                ParseInt::default(),
-                ParseString::default(),
-                Forward::default(),
-                Var::default(),
-            ),
-            PowershellStrategy::default(),
-        )
-        .unwrap();
-
-        // We are waiting for
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap() // statement_list
-                .child(4)
-                .unwrap() // pipeline
-                .data()
-                .expect("Expecting inferred type"),
-            Raw(Str(String::from("121212")))
-        );
-    }
-    #[test]
-    fn test_mul_assignment_operator_str_str() {
-        // infer global var in function statement
-        let mut tree = build_powershell_tree("$a=\"12\";$a*=\"3\";$a").unwrap();
-
-        tree.apply_mut_with_strategy(
-            &mut (
-                ParseInt::default(),
-                ParseString::default(),
-                Forward::default(),
-                Var::default(),
-            ),
-            PowershellStrategy::default(),
-        )
-        .unwrap();
-
-        // We are waiting for
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap() // statement_list
-                .child(4)
-                .unwrap() // pipeline
-                .data()
-                .expect("Expecting inferred type"),
-            Raw(Str(String::from("121212")))
-        );
-    }
-
-    #[test]
-    fn test_div_assignment_operator_int_int() {
-        // infer global var in function statement
-        let mut tree = build_powershell_tree("$a=10;$a/=2;$a").unwrap();
-
-        tree.apply_mut_with_strategy(
-            &mut (ParseInt::default(), Forward::default(), Var::default()),
-            PowershellStrategy::default(),
-        )
-        .unwrap();
-
-        // We are waiting for
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap() // statement_list
-                .child(4)
-                .unwrap() // pipeline
-                .data()
-                .expect("Expecting inferred type"),
-            Raw(Num(5))
-        );
-    }
-
-    #[test]
-    fn test_div_assignment_operator_int_str() {
-        // infer global var in function statement
-        let mut tree = build_powershell_tree("$a=10;$a/=\"2\";$a").unwrap();
-
-        tree.apply_mut_with_strategy(
-            &mut (
-                ParseString::default(),
-                ParseInt::default(),
-                Forward::default(),
-                Var::default(),
-            ),
-            PowershellStrategy::default(),
-        )
-        .unwrap();
-
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap() // statement_list
-                .child(4)
-                .unwrap() // pipeline
-                .data()
-                .expect("Expecting inferred type"),
-            Raw(Num(5))
-        );
-    }
-    #[test]
-    fn test_div_assignment_operator_str_int() {
-        // infer global var in function statement
-        let mut tree = build_powershell_tree("$a=\"10\";$a/=2;$a").unwrap();
-
-        tree.apply_mut_with_strategy(
-            &mut (
-                ParseInt::default(),
-                ParseString::default(),
-                Forward::default(),
-                Var::default(),
-            ),
-            PowershellStrategy::default(),
-        )
-        .unwrap();
-
-        // We are waiting for
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap() // statement_list
-                .child(4)
-                .unwrap() // pipeline
-                .data()
-                .expect("Expecting inferred type"),
-            Raw(Num(5))
-        );
-    }
-    #[test]
-    fn test_div_assignment_operator_str_str() {
-        // infer global var in function statement
-        let mut tree = build_powershell_tree("$a=\"10\";$a/=\"2\";$a").unwrap();
-
-        tree.apply_mut_with_strategy(
-            &mut (
-                ParseInt::default(),
-                ParseString::default(),
-                Forward::default(),
-                Var::default(),
-            ),
-            PowershellStrategy::default(),
-        )
-        .unwrap();
-
-        // We are waiting for
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap() // statement_list
-                .child(4)
-                .unwrap() // pipeline
-                .data()
-                .expect("Expecting inferred type"),
-            Raw(Num(5))
-        );
-    }
-
-    #[test]
-    fn test_mod_assignment_operator_int_int() {
-        // infer global var in function statement
-        let mut tree = build_powershell_tree("$a=9;$a%=2;$a").unwrap();
-
-        tree.apply_mut_with_strategy(
-            &mut (ParseInt::default(), Forward::default(), Var::default()),
-            PowershellStrategy::default(),
-        )
-        .unwrap();
-
-        // We are waiting for
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap() // statement_list
-                .child(4)
-                .unwrap() // pipeline
-                .data()
-                .expect("Expecting inferred type"),
-            Raw(Num(1))
-        );
-    }
-
-    #[test]
-    fn test_mod_assignment_operator_int_str() {
-        // infer global var in function statement
-        let mut tree = build_powershell_tree("$a=9;$a%=\"2\";$a").unwrap();
-
-        tree.apply_mut_with_strategy(
-            &mut (
-                ParseInt::default(),
-                ParseString::default(),
-                Forward::default(),
-                Var::default(),
-            ),
-            PowershellStrategy::default(),
-        )
-        .unwrap();
-
-        // We are waiting for
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap() // statement_list
-                .child(4)
-                .unwrap() // pipeline
-                .data()
-                .expect("Expecting inferred type"),
-            Raw(Num(1))
-        );
-    }
-    #[test]
-    fn test_mod_assignment_operator_str_int() {
-        // infer global var in function statement
-        let mut tree = build_powershell_tree("$a=\"9\";$a%=2;$a").unwrap();
-
-        tree.apply_mut_with_strategy(
-            &mut (
-                ParseInt::default(),
-                ParseString::default(),
-                Forward::default(),
-                Var::default(),
-            ),
-            PowershellStrategy::default(),
-        )
-        .unwrap();
-
-        // We are waiting for
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap() // statement_list
-                .child(4)
-                .unwrap() // pipeline
-                .data()
-                .expect("Expecting inferred type"),
-            Raw(Num(1))
-        );
-    }
-    #[test]
-    fn test_mod_assignment_operator_str_str() {
-        // infer global var in function statement
-        let mut tree = build_powershell_tree("$a=\"9\";$a%=\"2\";$a").unwrap();
-
-        tree.apply_mut_with_strategy(
-            &mut (
-                ParseInt::default(),
-                ParseString::default(),
-                Forward::default(),
-                Var::default(),
-            ),
-            PowershellStrategy::default(),
-        )
-        .unwrap();
-
-        // We are waiting for
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap() // statement_list
-                .child(4)
-                .unwrap() // pipeline
-                .data()
-                .expect("Expecting inferred type"),
-            Raw(Num(1))
-        );
-    }
-
-    #[test]
-    fn test_infer_local_var_type() {
-        let mut tree = build_powershell_tree("try{$foo = 1;$foo + 2}catch{}").unwrap();
-
-        tree.apply_mut_with_strategy(
-            &mut (
-                ParseInt::default(),
-                AddInt::default(),
-                ParseString::default(),
-                Forward::default(),
-                Var::default(),
-            ),
-            PowershellStrategy::default(),
-        )
-        .unwrap();
-
-        // We are waiting for
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap() // statement_list
-                .child(0)
-                .unwrap() // try_statement
-                .child(1)
-                .unwrap() // statement block
-                .child(1)
-                .unwrap() //statement list
-                .child(2)
-                .unwrap() //pipeline
-                .data()
-                .expect("Expecting inferred type"),
-            Raw(Num(3))
-        );
-    }
-
-    #[test]
-    fn test_array_concatenation() {
-        let mut tree = build_powershell_tree(
-            r#"
-            $foo = 1, 2
-            $bar = 3, 4
-            $foo += $bar
-            Write-Host $foo
-            "#,
-        )
-        .expect("A valid powershell tree");
-
-        tree.apply_mut_with_strategy(
-            &mut (
-                ParseInt::default(),
-                AddInt::default(),
-                ParseString::default(),
-                Forward::default(),
-                ParseArrayLiteral::default(),
-                Var::default(),
-            ),
-            PowershellStrategy::default(),
-        )
-        .unwrap();
-
-        // We are waiting for
-        assert_eq!(
-            *tree
-                .root()
-                .unwrap()
-                .child(0)
-                .unwrap() // statement_list
-                .child(3)
-                .unwrap() // Write-Host pipeline
-                .child(0)
-                .unwrap() // Write-Host pipeline_chain
-                .child(0)
-                .unwrap() // Write-host command
-                .named_child("command_elements")
-                .unwrap() // cmd elements
-                .child(1)
-                .unwrap() // args
-                .data()
-                .expect("Expecting inferred type"),
-            Array(vec![Num(1), Num(2), Num(3), Num(4)])
-        );
-    }
-}
